Move ForwardServer into its own files.
